Remove redundant exception handling

+++ gofit-app/gofit-app/src/main/java/com/gofit/service/UserServiceImpl.java
@@ -5,16 +5,20 @@ import com.gofit.entity.Roles;
 import com.gofit.entity.User;
 import com.gofit.dao.UserDAO;
 import com.gofit.exception.ResourceNotFound;
+import jakarta.persistence.TypedQuery;
 import jakarta.transaction.Transactional;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@Validated
 public class UserServiceImpl implements UserService {
 
     private final PasswordEncoder passwordEncoder;
@@ -30,23 +34,28 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public User save(User user) {
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
+    public User save(@Valid User user) {
+        user.setPassword(passwordEncoder.encode(user.getPlainPassword()));
         userDAO.save(user);
         return user;
     }
 
     @Override
     @Transactional
-    public User update(User user) {
-        User newUser = userDAO.update(user);
-        return newUser;
+    public User update(@Valid User user) {
+        Optional<User> optionalUser = Optional.ofNullable(userDAO.findByID(user.getId()));
+        optionalUser.orElseThrow(() -> new ResourceNotFound("User Not Found"));
+        user.setPassword(passwordEncoder.encode(user.getPlainPassword()));
+        return userDAO.update(user);
     }
 
     @Override
-    public Optional<User> findByID(int id) {
+    public User findByID(int id) {
         User user = userDAO.findByID(id);
-        return Optional.ofNullable(user);
+        if (user == null) {
+            throw new ResourceNotFound("User Not Found");
+        }
+        return user;
     }
 
     @Override
@@ -55,44 +64,49 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Optional<User> findByUsername(String username) {
+    public User findByUsername(String username) {
         User user = userDAO.findByUsername(username);
-        return Optional.ofNullable(user);
+        if (user == null) {
+            throw new ResourceNotFound("User Not Found");
+        }
+        return user;
     }
 
     @Override
     @Transactional
     public void addRoleToUser(int userID, int roleID) {
-        User user = findByID(userID).orElseThrow(() -> new ResourceNotFound("User not found with id " + userID));
+        User user = findByID(userID);
         List<Roles> roles = user.getRoles();
         if (roles == null) {
             roles = new ArrayList<Roles>();
         }
-        Roles role = rolesService.findById(roleID).orElseThrow(()
-                -> new ResourceNotFound("Role not found with id " + roleID));
+        Roles role = rolesService.findById(roleID);
         roles.add(role);
         user.setRoles(roles);
-        userDAO.save(user);
+        userDAO.update(user);
     }
 
     @Override
     @Transactional
     public void deleteRoleFromUser(int userID, int roleID) {
-        User user = findByID(userID).orElseThrow(() -> new ResourceNotFound("User not found with id " + userID));
+        User user = findByID(userID);
         List<Roles> roles = user.getRoles();
-        if (roles == null) {
+        Roles role = rolesService.findById(roleID);
+        if (roles == null || !roles.contains(role)) {
             return;
         }
-        Roles role = rolesService.findById(roleID).orElseThrow(()
-                -> new ResourceNotFound("Role not found with id " + roleID));
         roles.remove(role);
         user.setRoles(roles);
-        userDAO.save(user);
+        userDAO.update(user);
     }
 
     @Override
     @Transactional
     public void deleteByID(int id) {
+        if (userDAO.findByID(id) == null) {
+            throw new ResourceNotFound("User not found");
+        }
         userDAO.delete(id);
     }
+
 }
